security-package/issues/106: Cover ReCaptchaCustomer module with integration tests - update.
